requirements.txt
yfinance==0.2.28
pandas==2.2.2
numpy==1.26.4
pandas-ta==0.3.14b0
supabase==2.3.4
google-api-python-client==2.149.0
xai-sdk
ib_insync==0.9.86
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.19.0
asyncio-mqtt==0.16.2
pydantic==2.6.4
fastapi==0.110.1
uvicorn==0.29.0
